make general functions
